Fixed sign onversions in JxHash and JxRuntime
In JxHash cast the len to unsigned int before xor with seed.
Since it is used as its bit value and not the actual length,
that cast is safe.
I used a C-Style cast here to be consistent within the file.

In JxRuntime initialize local offset as unsigned int 0u.
as it is used for calculation with unsigned field symbol->columnNumber.

+++ Source/JxHash.cpp
@@ -15,7 +15,7 @@ namespace Jinx::Impl
 		const uint32_t m = 0x5bd1e995;
 		const int r = 24;
 
-		uint32_t h = seed ^ len;
+		uint32_t h = seed ^ (unsigned int)len;
 
 		const unsigned char * data = (const unsigned char *)key;
